pyproject.toml
[tool.poetry]
name = "tech-jobs-data-platform"
version = "0.1.0"
description = ""
authors = ["dmaruyama-51 <0501awasuram@gmail.com>"]
readme = "README.md"
package-mode = false  # パッケージモードを無効化

[tool.poetry.dependencies]
python = "^3.11"
functions-framework = "^3.8.2"
google-cloud-storage = "^2.19.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
google-cloud-logging = "^3.11.3"
opentelemetry-api = "^1.29.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
mypy = "^1.14.1"
ipykernel = "^6.29.5"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
sqlfluff = "^3.3.0"
sqlfluff-templater-dbt = "^3.3.0"


[tool.poetry.group.scraper.dependencies]
bs4 = "^0.0.2"


[tool.poetry.group.loader.dependencies]
google-cloud-bigquery = "^3.27.0"


[tool.poetry.group.dbt.dependencies]
dbt-core = "^1.9.1"
dbt-bigquery = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "opentelemetry.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.*"
ignore_missing_imports = true
